add https://github.com/oerdnj/deb.sury.org/issues/1768
The owners of the deb.sury.org repository, the maintainers of the PHP packages for Debian and Ubuntu, turned out to be hostile to IPv6 users who wanted to get the PHP packages. The owners still don't know how to configure a proper DNS record in the current year and plan to blocklist the IPv6 users trying to use their PHP packages, making it inconvenient for many IPv6 users to get PHP packages for Debian and Ubuntu.