New Fish-shell drama by Rustacean: Ridiculousfish
New Fish-shell drama by Rustacean: Ridiculousfish

Fish-shell is a user-friendly command line shell for Linux, macOS, and the rest of the family. A GitHub user named Ridiculousfish is a member of the Fish-shell organization and an Apple software engineer. 

Ridiculousfish opened a pull request titled "Rewrite it in Rust" in January 2023, the URL of which was shared on Y Combinator Hacker News and prompted numerous off-topic comments from Rust missionaries and Rustaceans who knew nothing about Fish-shell.

Primary sources and web archives:

https://github.com/fish-shell/fish-shell/pull/9512
https://web.archive.org/web/20230131070907/https://github.com/fish-shell/fish-shell/pull/9512
https://archive.ph/j5Y3q